requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.0
simple-salesforce==1.12.4
imaplib2==3.6
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
schedule==1.2.0
logging==0.4.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
